// projects/bible/web/src/interfaces/bibles/versesChapterOfBook.ts
// Generated by https://quicktype.io
import { Version } from './versionesResponses';

export interface VersesChapterOfBook {
  ok: boolean;
  results: number;
  status: boolean;
  code: number;
  version: Version;
  verses: Verse[];
}

export interface Verse {
  book_number: number;
  chapter: number;
  verse: number;
  text: string;
}