// projects/bible/web/src/interfaces/bibles/allBooksResponses.ts
// Generated by https://quicktype.io

export interface BooksResponses {
  ok: boolean;
  results: number;
  status: boolean;
  code: number;
  version: Version;
  books: Book[];
}

export interface Book {
  book_number: number;
  short_name: string;
  long_name: string;
  chapters: number;
  verses: number;
}

export interface Version {
  language: string;
  longLanguage: string;
  name: string;
  description: string;
  file: string;
}